Move topic related files to the pages folder
