+++ src/App.js
@@ -2,8 +2,7 @@ import React, { Component } from 'react'
 import { Redirect, Route, Switch } from 'react-router-dom'
 import { Container } from 'semantic-ui-react'
 import { graphql } from 'react-apollo'
-import { Home, LeanCoffeeDetails, ListLeanCoffees, ListParticipants } from './pages'
-import ListTopics from './components/topics/List'
+import { Home, LeanCoffeeDetails, ListLeanCoffees, ListParticipants, ListTopics } from './pages'
 import Loading from './components/loading/Loading'
 import Navbar from './components/menu/Navbar'
 import CreateUser from './components/user/CreateUser'
