+++ src/pages/index.js
@@ -4,3 +4,5 @@ export { default as LeanCoffeeDetails } from './coffee/details/Details'
 export { default as Home } from './home/Home'
 
 export { default as ListParticipants } from './participant/List'
+
+export { default as ListTopics } from './topic/list/List'
